requirements.txt
einops==0.6.1
matplotlib==3.7.2
numpy==1.24.4
pandas==2.0.2
Pillow==10.0.0
torch==1.13.1+cu116
torchvision==0.14.1+cu116
wandb==0.15.4